+++ CHANGELOG.md
@@ -10,7 +10,11 @@ Types of changes
     Security in case of vulnerabilities.
 )
 
-## [2.0.1](https://github.com/contao-themes-net/mate-theme-bundle/tree/2.0.1) – 2022-11-04
+## [2.1.0](https://github.com/contao-themes-net/nature-theme-bundle/tree/2.1.0) – 2022-11-30
+
+- [Changed] Add multi domain support
+
+## [2.0.1](https://github.com/contao-themes-net/nature-theme-bundle/tree/2.0.1) – 2022-11-04
 
 - [Changed] Change migration order
 
@@ -20,6 +24,10 @@ Types of changes
 - [Added] Add migrations for demo data import (Setup without further steps, install and run migrations -> Done!)
 - [Removed] Cleanup older Contao SQL files
 - [Added] Add more options to style content slider
+
+## [1.10.0](https://github.com/contao-themes-net/nature-theme-bundle/tree/1.10.0) – 2022-11-30
+
+- [Changed] Add multi domain support
 
 ## [1.9.2](https://github.com/contao-themes-net/nature-theme-bundle/tree/1.9.2) – 2022-10-11